Add semantic coloring, updates based on feedback
This is a combined updated, based on the feedback I got from several
small sessions with one or two people.  Major update is semantic
coloring - it is especially helpful when the same name is used for both
a type and a constructor.

Update the color scheme - I think the new one looks better.  It is based
on the Emacs Haskell color scheme.

Removed placeholder image references.  Should someone add new images,
they would need to be positioned from scratch, and may be positioned in
different slides.

Keep Haskell logo both in PNG and SVG.  SVG is the original.

Some of the examples updated based on the feedback.  Added function and
operator type definitions to slides where they are referenced.  This
help new people make the connection, not requiring them to remember type
definitions provided one or two slides ago.  Some cases are still not
covered.

Added references to two new articles that I think could be interesting
to new people (one for each of 101 and 102).