Remove CODELAB define, hlint suggestions
There seems to be no reason to use both "CODELAB" and "SOLUTION"
defines.  One is enough.  hlint is skipping all the conditional defines,
as there is no cabal file, and as a consequence, it is showing a lot of
undefined functions.

Removing "CODELAB" makes it a bit easier for the attendies, as they can
load the codelab code in ghci with one less command.