101 slides: Add sections 1 and 2
There was a general complain, that it was too hard to listen to such a
long theoretical discussion before we do anything.  Two new codelab
sections plus some material reordering should make it possible to have
two short codelab sessions in the middle of the talk.  We will still
have a larger codelab part at the end, after people have been introduced
to a larger number of different constructs.