Recommend why functional programming matters
Want to recommend the article why functional programming matters.

This is not a scared research paper but rather a modest introduction on how functional programming plays an important role in software development using many interesting examples. It is a quite famous paper that got more than 1k citations and is usually considered a good article to read when understanding the motivation of functional programming.

I checked adding this reference still fits the page.